Update patch set 5
Patch Set 5:

Had to change the script name because the one I used is the one that the script executes by default. As such we need to use a different name so we can execute the test in a different phase.

Patch-set: 5
Reviewer: Gerrit User 5473 <5473@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0