jjb: xci: xci-verify-jobs: Fix clean VM job
According to
https://docs.openstack.org/infra/jenkins-job-builder/definition.html#macro-notes
we only need to use ${{x}} when the macro has parameters. However, none
of our macros has so the correct way to use the shell variable is ${x}.
Moreover, we set the missing execute bit in the jenkins test script.
Additionally, we fix the OPNFV_RELENG_DEV_PATH parameter to point to the
directory inside the VM where the releng-xci repo is located.
Finally we fix the parent directory in which the test is executed.

Change-Id: Ic6157caa5042274a3a3d0de276bdacc15a90a82d
Signed-off-by: Markos Chandras <mchandras@suse.de>