Update patch set 4
Patch Set 4:

> We don't know if docker is there, so we make sure.
 > 
We could definitely do a better job at documenting what's on the build servers. :( Docker is on all the servers labeled 'opnfv-build', and one other issue is that two of those servers are Ubuntu. So any yum commands will fail. If you want to restrict this job to CentOS you can use the parameter 'opnfv-build-centos-defaults'.

 > Sleep a bit , wait till all 7 containers runs, before functional
 > test

Ah that make sense. Thanks for the clarification!

Patch-set: 4