Update patch set 3
Patch Set 3:

For this release we're going with ubuntu. I tested the alpine containers and we have problems with restapi and vnf. The rest of the images build OK with alpine/multiarch as base image for functest-core. Maybe we can work together with what we're doing with manifests and see what is to be done with arm. Is manifest based on images with architecture set as tag?

Patch-set: 3