Update patch set 2
Patch Set 2:

> Let's wait until we decide if we leave OSA or osa

This is really a code change right? You rename the module from OSA to osa so from the python point of view I think it's better. I was just pointing out that OSA is fine if you want to refer to the project in docs or something.

Patch-set: 2