fetch_os_creds.sh: Fix unbound BUILD_TAG
Fuel baremetal deployments provide a certificate, which should be
copied over to test containers. Virtual deployments do not, hence
the need to determine the deployment type and decide whether the
certificate should be transferred in the first place.

However, the Docker args we contruct will try to map the certificate
file no matter what, so we will instead try to transfer the cert
from Salt Master in all deployments, and just create an empty file
in case it does not exist.

Change-Id: I3d3fe6cabf552a88841c529ce38446b2e54b3053
Signed-off-by: Alexandru Avadanii <Alexandru.Avadanii@enea.com>