Update patch set 1
Patch Set 1:

> what was the previous high score? I'm missing how we are checking
 > if 3 or greater is 30, and each iteration goes up by 10, but we
 > were calculating before on 4 iterations.  In my mind top score
 > should have been 40?

Sorry my comment is misleading (I should never commit watching TV on friday night)

the score of a scenario is the addition of the test case scores (without weight, so far connectivity = refstack)

each case can get 3 points (last 4 iterations PASS), 2 points (more than 1 PASS over the last 10 days but at least 1 of the last 4 is FAIL), 1 point (at least 1 PASS over the last 10 days) or no point (never work)

so if we consider a scenario with 10 cases, the top score is well 30 = 10*3 (max points)

the coefficient to calculate the top score is not the number of iterations

1 iteration => each test case can only get 1 point => possible  top score 10
2 iterations  => top score 20
3 iterations => top score 20
4 iterations => top score 30
+ iterations => top score 30

note that only tests from healthcheck or smoke tiers are used for calculation (feature tests which were very different from one installer to another are not taken into account for the calculation)

https://wiki.opnfv.org/plugins/servlet/mobile#content/view/6828617

the calculaton shall be reviewed after Euphrates
weight on the test complexity shall be probably added.

It shall be discussed in the testing group.

Patch-set: 1
Label: Verified=0