KVMFORNFV: Disable jobs on Euphrates
This patch is used to diasable jobs on Euphrates branch as there is only
one node assigned for the daily job. If the jobs on master and Euphrates
are enabled, the verify job is delayed.So inorder to avoid the delay for
the verify job, disabling the jobs on Euphrates.

Change-Id: Ib0eff4a1156858a73e4b889a61ac0cbd4ebe5ef0
Signed-off-by: Navya Bathula<navyax.bathula@intel.com>