Update patch set 1
Patch Set 1:

> and if you want to overwritte it (for use of a local DB - I assume
 > it is the need behind, no?) you still can use an env variable(from
 > your local jenkins) but no need in sh scripts for CI

yes, this patch aims to use the env variable to rewrite the TESTAPI_URL from local Jenkins or start a Functest testing manually. If only provide the url, nothing change here, the URLs leveraged in containers will always be testresults.opnfv.org, because TEST_DB_URL only is changed with -e option when docker run the functest-xxx containers, or else we need to start the container first, go inside the container and change it, or docker cp the file out of the container -> edit it -> copy it back and restart the container, all include lots of steps and wasting time
Considering we have multiple images now, the operation will multi-fold as well. and considering the upcoming Plugfest, so many pods wait for testing.
So I suggest either we change here or we add a one-click-run tool(script) under Functest repo, what do you think?

Patch-set: 1