Build Docker images with jjb
WORK IN PROGRESS Multijob for building the docker images

Beside the images that have dependencies ( as functest-components,
functest-helthcheck, functest-restapi, functest-features,
functest-smoke, functest-vnf, functest-promiser) all docker images
are build SEQUENTIALY in a multijob as Phase1.
The images with dependencies are build in PARALLEL as Phase2
only is the first Phase result is SUCCESFULL.
The manifest is created by opnfv-docker.sh as final job,
only for specific dockerrepos.

Change-Id: Iaa6640568ed18a3b087db56449173841dd2757ff
Signed-off-by: Delia Popescu <delia.popescu@enea.com>