[docs] uploading RTD build results for preview
Currently, there are two toolchains being used for docs: RTD and a
simple preview build. The latter are used as previews in Gerrit whereas
the former is the official documentation. Both toolchains generate
different results.

The docs-verify-rtd job creates a full build of the documentation just
like RTD does. This patchset replaces the upload of the preview build
with the docs build by the docs-verify job.

Change-Id: I45d9dd41fb417d4720537049a8954378e8764b36