Update patch set 5
Patch Set 5:

Thanks Aric. This tool chain mismatch caused quite some confusion in certain projects.

One comment: I am not yet convinced that compiling the whole documentation into a single file helps devs during the review process. It seems cumbersome to find the right piece of documentation if you are just working on one particular file.

In a patch [1] I pushed yesterday (not knowing about this one here yet), I attempted to upload either the entire documentation to artifacts if the patchset is for the opnfvdocs project, or just the docs generated in the submodule if a patchset is for a non-opnfvdocs project.

[1] https://gerrit.opnfv.org/gerrit/#/c/46173

Patch-set: 5
Label: Verified=0