Update patch set 18
Patch Set 18:

(3 comments)

I think I need some clarification as written inline.

The commit message says "The images with dependencies are build in PARALLEL in the second Phase". Is this done within the execute shell part of that build-child-img phase or are we supposed to have multiple jobs (one job per child image) running in parallel as part of build-child-img phase.

Patch-set: 18
Label: Verified=0