Update patch set 7
Patch Set 7:

It feels a bit risky to delete the old artifacts build if something would go wrong with readthedocs... You don't see a need to keep it but not publish it? Or maybe that doesn't make sense since the build as you say is different from the one that we have today.

Patch-set: 7
Label: Verified=0