Update patch set 19
Patch Set 19:

I should have precised that we can't manage functest-parser containers (Euphrates and master) in the same way. As parser requires Pike even for E-release , we can't build the related E container on top of functest-core. For master, it will be simpler when we upgrade Functest to Pike too (following OPNFV rules).

I think it's fine to prepare the manifest for master only.

Patch-set: 19