Update patch set 26
Patch Set 26:

Alec,

> If I understand correctly: if there are N projects that build docker images we first build all these N images in sequence.
> And only if all N images succeed the second batch of functest child images are built?
> Or is it finer grain - at the level of every repo (meaning storperf and yardstick containers can be built concurrently)?
This patch changes things only for Functest due to Functest having the need to build core first followed by child images, building concurrently.
So, other projects still continue building things in same way; single job, building and pushing images.
https://gerrit.opnfv.org/gerrit/gitweb?p=releng.git;a=blob;f=jjb/releng/opnfv-docker.yml;hb=HEAD#l185
Depending on what a certain project needs, different build processes can be employed.

> Which repos get their images rebuilt?
All the projects that are added to opnfv-docker.yml get their builds triggered based on poll scm.
https://gerrit.opnfv.org/gerrit/gitweb?p=releng.git;a=blob;f=jjb/releng/opnfv-docker.yml;hb=HEAD#l41

> Only those that have a changeset submitted in gerrit? On what trigger? 
As far as I know, we do not build & push based on patchset-submitted. 
We do build stuff for Storperf but those images are never pushed as the purpose of this build is to provide feedback from builds to ensure patches don't break things.
Again this is very project specific so if you want to push images once a change gets merged and skip pollscm, it can be done.

> Just too many questions maybe we need a short document that describes how all this works (gerrit not best format). We need to document clearly: how builds are trigerred in what order/sequence/concurrency
This and more will be documented on docs.opnfv.org. We simply lack human resources to do everything altogether so docs get a bit downprioritized.

> As of today it looks like the builds are triggered periodically (even when there is no change in repo) on an unknown period (looks like once or twice a day).
We use polling so the builds are not triggered if there is no change in repo.
https://docs.openstack.org/infra/jenkins-job-builder/triggers.html#triggers.pollscm
https://build.opnfv.org/ci/view/functest/job/functest-docker-build-push-euphrates/scmPollLog/

> Also would be nice to mention how to trigger a container build based on tag (got a piece of it in another review but missing the other half).
Same as the other documentation question; we will do this once we fix the jobs.

Patch-set: 26