fix the bug of patch verify
If one patch don't modify the "yml" or "yaml" file,
the script will exit with 1 at egrep "ya?ml$", then mark build as failure.
In my opinion, if not modify the "yml" or "yaml" file, the yamllint should be successful.
So we need mark the result of egrep "ya?ml$" as true, then judge the $YAML_FILES.

Change-Id: I8abcadc9e4c85935ac7b20cdf54713b9800108ba
Signed-off-by: wutianwei <wutianwei1@huawei.com>