Update patch set 2
Patch Set 2: Code-Review-2

(1 comment)

I don't see the rational for this.

If YAML_FILES is empty, which it will be if a patch doesn't contain yaml files given the job is triggered by changes to yaml files, the next statement 'for file in YAML_FILES' won't execute, and the job will complete successfully.

Patch-set: 2