jjb: xci: bifrost-verify: Simplify bifrost job
We no longer test OPNFV commits that affect bifrost since these are
being tested as part of the XCI jobs. As such, simplify the job to
only test incoming upstream patchsets. Moreover, we need to block
when an XCI job is running since both jobs run on the same hosts and
create the same VMs.

Change-Id: I5c11ae5b6a4d719ec41a6e663ddf6ecbd71ac1a5
Signed-off-by: Markos Chandras <mchandras@suse.de>