Converting INFO file to Yaml
This is a proposed format for switching the INFO file to a
machine readable format.

Some details may be wrong: notably Linux Foundation IDs and timezones.
These details aren't relevant to the discussion of the overall format
and structure of the file, and can be updated later.

Notable changes to the INFO file:

 * Adds Meeting information: time, date, location, agenda, etc.
 * Updates Commiter info to include: Linux Foundation IDs and Timezone

Change-Id: I05aaef36e44978d2abc73f5f18757d91ee6002d0
Signed-off-by: Trevor Bramwell <tbramwell@linuxfoundation.org>