xci: Create periodic jobs for upstream osa
This change takes xci-verify-jobs as base and adjusts osa-periodic
jobs to ensure they do what they are supposed to.

The structure of periodic jobs is pretty similar to xci-verify-jobs and the
differences comparing to them are
- OPENSTACK_OSA_VERSION is set to master.
- SCM type changed to git-scm from gerrit
- Voting is removed.
- Triggering will be done using pollscm to ensure we only run jobs when something
changes upstream. Triggers are not enabled yet until the basics are verified
working.

This change requires https://gerrit.opnfv.org/gerrit/#/c/49275/ in order for
periodic jobs to work as expected.

Further changes are needed to finalize the structure.

Change-Id: I8f6519ccce93e6b8b65509638294419459fd36e8
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>