RFC: Set the basic structure for post merge jobs
This change proposes structure for  post-merge jobs for the
scenarios that are onboarded to XCI.

The idea with these jobs is to verify the sanity of impacted
scenario to find out if the version that is tested should be
promoted to the next level in XCI; daily on baremetal.

If it succeeds, the metadata regarding the scenario needs to
be stored somewhere (artifact repo, etc.) so daily jobs can
poll that location to see if there is any scenario that is
worth testing on baremetal.

If it fails, the scenario could be demoted instead which still
requires further discussion and thinking. We might perhaps leave
the metadata of the version of the scenario with previous sha
as the candidate for testing instead of demoting it since we
need to have working version of the scenarios at all times.

Change-Id: I85d7176b16d0d6da45cd2462b0cdb990485b6b0d
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>