jjb: xci: xci-verify-jobs: Do not trigger on forbidden paths
We should set 'disable-strict-forbidden-file-verification' to 'false'
to really prevent any changes to the forbidden paths to trigger a CI
job. When this option is set to 'True', then a patchset that modifies
both allowed and forbidden paths will trigger a job. However, in order
to be on the same side in regards to the start-new-vm.sh script, we
really shouldn't trigger the job. Moreover, we drop the rest of the
forbidden paths since we can use the 'skip-verify' topic name to
skip the deployment.

Change-Id: I2a89e51861c7fb8541fe774b128ea4049e91df43
Signed-off-by: Markos Chandras <mchandras@suse.de>