Create functest job to run on itri-pod1
Joid has been installed on ITRI POD manually and functest job is
needed to run testing on it.

This change creates the job but since no triggers are defined for
standalone functest jobs, the pod owner might need job build rights
which can be discussed as part of the review of this change.

It uses Euphrates version of functest.

Change-Id: I8698964464d25059709c3e4826e48645a3484588
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>