Update patch set 1
Patch Set 1: Code-Review-1

I know the current aarch64 docker image for dovetail is opnfv/dovetail_aarch64 but I would rather change that to have the opnfv/dovetail repo with x86_64 and aarch64 tags. This is the approach storeperf and functest have and it's what we agreed going further for docker images.

Patch-set: 1