Update patch set 1
Patch Set 1: Code-Review+1

> I know the current aarch64 docker image for dovetail is
 > opnfv/dovetail_aarch64 but I would rather change that to have the
 > opnfv/dovetail repo with x86_64 and aarch64 tags. This is the
 > approach storeperf and functest have and it's what we agreed going
 > further for docker images.

I agree, going forward we should refactor the docker image build using manifests and drop the '<repo>_<arch>' scheme.
But just to be able to debug the docker image in parallel with that refactor, my opinion is that we can merge this as-is for now.

Patch-set: 1
Reviewer: Gerrit User 2997 <2997@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=+1