Update patch set 3
Patch Set 3:

> please move deployment relevant changes to the releng-utils repo

I am afraid that these deployment scripts can't be moved to utils. Because they are a relevant independent module to get the infomation of the SUT. Except daisy, there are already compass/adapter.py, apex/adapter.py, fuel/adapter.py and osa/adapter.py.
It needs to move all modules folder to utils if we really want to do this.

Patch-set: 3