Update patch set 4
Patch Set 4:

(1 comment)

More of a general comment, it would be great to also add somewhere (README or comment) how project owners should use the feature (not everyone is expert in reading jjb files).
 It is unclear for example to see at what stage of gerrit review, the docker build is triggered: is it after the diff is +2 and merged? or before?
Can for example the docker build be triggered before the merge (in the Verify stage) and if it fails, fail the Verify. If success, the docker image would then be pushed to dockerhub only after te code is +2 and merged?

Patch-set: 4
Label: Code-Review=0