Reduce Basic Job Duplication
In the initial formation of projects we found it easy to copy job
definitions to provide project a baseline job config to work off. This
has led to a lot of duplication and misalignment with the gerrit
triggers and default build server tag to be used.

Collapsing these jobs into a job-group containing the stream and a
verify job per-stream, should help reduce this duplication.

Change-Id: Icb366487590a145be4cbfc0637a8d86a6d9b7cec
Signed-off-by: Trevor Bramwell <tbramwell@linuxfoundation.org>