merge lint builders to 'lint-diff-code'
There are multiple builders for lint and syntax check. Those are
definied in different ways. OPNFV has common coding style [1].
So, all repos should be checked by one generic checker builder/job.

'lint-yaml-code' builder is cool as it only check modified files.
This won't require immediate fixes of violations in whole repo code
right after the lint checker job is added.

This patch replaces all existing lint bulders by the new builder
'lint-diff-code'. This checkes codes in bash, python and yaml files
modified by the new patchset. If there is any violation, it will be
reported to the gerrit page of the patchset.

This patch also fixes the unexpected override of yaml config. The
default value of parameter 'allow-non-breakable-words' is true, but
it was set false since the item in dict 'line-length' is missing.

Note that lint jobs are still configured to run against some project
repo, in order to seperate discussion; wheter we should apply this
lint job to all OPNFV repo or not immediately.

[1] https://wiki.opnfv.org/display/DEV/Contribution+Guidelines

Change-Id: Ib4ab3ba5dc08845f3016b8be772f4ed119f11b2d
Signed-off-by: Ryota MIBU <r-mibu@cq.jp.nec.com>