merge lint builders to 'lint-diff-code'
There are multiple builders for lint and bash syntax check.
'lint-yaml-code' builder is cool as it only check modified files.
This won't require immidiate violation fix during the process of
adding lint jobs.

The aim of this patch is to have single common lint builder which can
be applied to all OPNFV repos in order to keep better readability of
codes, and to follow the idea of the diff checking in 'lint-yaml-code'.
Also, merging check builder and report builder into one can avoid
mistake in new jjb configuration.

This patch replace all those lint bulders by 'lint-diff-code' in which
codes in modified bash, python and yaml files will be checked, and all
violation will be posted to gerrit page of the patchset as a comment
if there is any violation.

Lint jobs are still configured to run against some project repo, in
order to seperate discussion; wheter we should apply this lint job
to all OPNFV repo or not.

Change-Id: Ib4ab3ba5dc08845f3016b8be772f4ed119f11b2d
Signed-off-by: Ryota MIBU <r-mibu@cq.jp.nec.com>