Add GERRIT_REFSPEC to 'project-parameter' Macro
GERRIT_REFSPEC is always passed by the gerrit-trigger plugin when a job
is triggered by Gerrit. Because it is not explicily defined, there is no
way to manually trigger jobs, as the git clone looks up the list of refs
by GERRIT_REFSPEC.

Being able to manually trigger jobs (with node parameters so they can be
restricted) is very helpful in debugging CI issues.

Change-Id: I8a1d9ea380902fc95f30482e5acb616347709ab1
Signed-off-by: Trevor Bramwell <tbramwell@linuxfoundation.org>