xci: Move start new vm execute shell to external script
The script used by verify jobs to create VMs for CI can be reused
by other jobs as well so it is moved out of the yaml file.

Apart from that, if we decide to change the way we run CI (not
using VMs for performance reasons), we don't have to make big changes
to yaml either.

Change-Id: If132a1ebe7c9c3d6999b98dc3d7e441694227306
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>