xci: Move new vm cleanup execute shell to external script
The script used by verify jobs to create VMs for CI can be reused
by other jobs as well so it is moved out of the yaml file.

Apart from that, if we decide to change the way we run CI (not
using VMs for performance reasons), we don't have to make big changes
to yaml either.

Change-Id: I40f9a1b912cd14a8b6485e13067f6965e2a4e126
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>