xci: Convert execute shell scripts to external scripts
The scripts used by verify jobs to create VMs, start deployments,
run functest, and do cleanup for CI can be reused by other jobs as
well so they are moved out of the yaml file.

Apart from that, if we decide to change the way we run CI (not
using VMs for performance reasons), we don't have to make big changes
to yaml either.

Change-Id: I2df1f2d60f999bcd7d53a5cd879112a5b2b0ca05
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>