xci sfc: Enable post merge deployment
This change enables the deployment upon merge of a change.
Followup patches will incorporate functest smoke and promotion
which the promotion will most likely be the first one as functest
parts are a bit tricky due to failures while running testing
in VMs.

Apart from the enablement of the deployments, job concurrency is
disabled as well since we should only test one change at a time
and attempt promotion for any scenario in order to ensure we keep
the promotion consistent.

Change-Id: Ib4b5754b6b88f42eab83c656e03773c83068016f
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>