Apex: Fixes scheduling with apex verify
apex-verify job calls apex-virtual to deploy and run functest.  Apex has
2 virtual slaves attached to the apex-virtual-master label.  When 2
verify jobs are triggered at the same time, apex-verify job is scheduled
on each slave correctly.  However, when the mulitjob triggers
apex-virtual jobs, it only schedules both of them to a single slave.

This happens even though apex-virtual job has the same slave label and
node parameters are not passed from verify job.  This patch changes the
node to be passed to apex-virtual from apex-verify.  That way 2 nodes
will be scheduled on, but this is still not ideal scheduling as more
than 1 verify can run per node (but not more than one virtual job).

Change-Id: I155351c9037f70df2c5dba11bb5592423849e760
Signed-off-by: Tim Rozet <trozet@redhat.com>