Include global-jjb v0.15.0
Global-JJB[1][2] is a collection of reusable JJB templates created by The
Linux Foundation Release Engineering team.

Using it will allow OPNFV to take advantage of the Release Engineering
work being done across all the other Linux Foundation Networking (LFN)
projects.

Just including is a safe as it will not use any of the jobs templates it
defines until they are listed in a jjb job definition.

[1] https://github.com/lfit/releng-global-jjb
[2] http://docs.releng.linuxfoundation.org/projects/global-jjb/en/latest/index.html

Change-Id: Id1548438b57df177a7a37cb3ed5de6ae0febab3d
Signed-off-by: Trevor Bramwell <tbramwell@linuxfoundation.org>