Enable verify jobs for external scenarios
This change extracts the scenario that is impacted by the patch
which triggered the verify jobs and saves the scenario name into
java properties file. This file will then be used by envInject
plugin to inject the scenario name into the build environment.

If the triggering change impacts multiple scenarios, the build
will result in  failure and the developer will be informed about
a possible way to rectify the issue.

Support for generic scenarios will be introduced in subsequent changes.

Change-Id: I6406299c1352b6e9b8a9297c5441adf03445be33
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>