jjb: xci: Add postbuild publisher script to remove clean VM
The clean VM normally stays around after the job is completed or killed
due to timeout. Unless a job for the same distro is scheduled on the
same node, that VM will state around for a while consuming valuable
system resources. As such, we should remove it after the job finishes
to free up resources for the new jobs.

Change-Id: I88b5e32797969d9378588199a7ddb013bbe740c9
Signed-off-by: Markos Chandras <mchandras@suse.de>