xci: Fix the Gerrit Trigger clone issue for external scenarios
When a patch comes in for an external scenario, Gerrit Trigger injects
refs for that patch for the corresponding project. With the previous
configuration, job attempts to get that patch for releng-xci instead
of the actual project.

But switching to the right project causes Jenkins jobs to fail since
the releng-xci will not be checked out to WORKSPACE.

This change makes the clone for the repo/patch work and then it wipes
out the WORKSPACE and clones releng-xci there to get the actual script
parts to work.

Change-Id: Ide31c21a209dff953118f51e9b09de1d206a3085
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>