Update patch set 3
Patch Set 3:

> I believe this is the right approach, but have some concerns.
 > 
 > Is there any process between two repo? For instance, can we get
 > "deprecated" macro/parameter info from the global repo? otherwise,
 > our jjb could easily broken.

I updated the commit message to explain this a bit more. I don't think there is a deprecation policy for global-jjb, but because the submodule is pointed at a tag, we need to manually update it to pull in the new version.

This is obviously something we would test first on the sandbox so any bugs in jobs added don't break our jobs.

Patch-set: 3
Reviewer: Gerrit User 1451 <1451@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0