xci: Set options in run-functest.sh
Functest healthcheck job actually fails but due to not setting
options properly in run-functest.sh, we failed to notice this.

Please note that this patch will enable changes to be blocked
since creation of external network and associated subnet fail,
due to an upstream issue. This causes functest healthcheck to
fail which in turn results in Verified-1 on Gerrit.

Change-Id: I32eeb422895930d039d3289ffb65e10763db9c4b
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>