xci: Set the installer and move scenario extraction into functions
The scenarios named os-* are deployed by osa and the ones
named k8-* are deployed by kubespray. This change determines
the installer to use by looking into the scenario name.

The installers are currently hardcoded within xci-set-scenario.sh
script and this needs to be fixed for other installers when they
onboard to xci.

Apart from setting the installer, extraction of generic and external
scenarios are moved into corresponding functions.

Change-Id: I19882bcd2a1260765e601ecbe517551c60401d74
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>