Define parameters for triggering jobs
In
https://wiki.jenkins.io/display/JENKINS/Plugins+affected+by+fix+for+SECURITY-170,
it indicates that:

    only build parameters that have been explicitly defined in a job's
    configuration will be available by default at build time. Any other
    arbitrary parameters added to a build by plugins will not be
    available by default.

It has been tested in sandbox.opnfv.org that only defined parameters can
be passed to the triggered jobs.

For parameters required by TESTAPI are common to all the projectes,
including installers, Functest, and Yardstick, it is convenient to
define a common macro and used by others.

Change-Id: Iac1da85445e1b5d75109ccd64b174f35151b9969
Signed-off-by: Julien <zhang.jun3g@zte.com.cn>