Update patch set 4
Patch Set 4: Code-Review-1

Thanks for creating this patch ahead of time!

With respect to how this interacts with the release automation work,
the automation will not attempt to add jobs if it already sees the fraser stream.

My suggestion is to add a dependent patchset that adds the stable branch for compass. I will include compass in the initial automation patch, so the file can just be copied to releases/fraser/compass.yaml.

When you're ready to branch, you can update the hash in the patchset, and we can merge this one first.

Patch-set: 4