Initial Release files from Euphrates Release
These files normally contain the start point of the stable/release branch,
but as this is the initial setup starting from the previous release,
they contain the tip of the stable branch.

When a project is ready to being the stable branch process, they will
copy their project's release file and modify the 'branches' mapping to
create their stable branch.

Change-Id: I0244da5eeb94c431d83f217ec07b08553fdf0639
Signed-off-by: Trevor Bramwell <tbramwell@linuxfoundation.org>