xci: Fix envInject and remove scenario distro support check
envInject doesn't play well with variable expansion so it is fixed. [1]
Scenario support check is not necessary for promotion since the
promotion job is common for all scenarios/distros and check is done
as part of earlier distro jobs accordingly. [2]
gsutil upload is also made quiet.

Fixed version can be seen from [3].

[1] https://build.opnfv.org/ci/job/xci-merge-promote-virtual-master/2/console
[2] https://build.opnfv.org/ci/job/xci-merge-promote-virtual-master/4/console
[3] https://build.opnfv.org/ci/job/xci-merge-promote-virtual-master/5/console

Change-Id: Ie36394cab631826209dda4cea6daa456aca3c276
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>