Update patch set 2
Patch Set 2:

Added others who are more involved in PDF/IDF work so they can shed some light to what I'm going to propose/ask.

Would it make sense for IDF to have this info; login info for all the nodes after based on how they get deployed by the installers? I don't exactly remember what should go into IDF and what shouldn't so if it is not there, Dovetail team can perhaps ask this to Infra/Pharos to see if the needs can be incorporated.

The reason for this is that PDF/IDF should be useful for more purposes and the more we use it, the less discrepancy around, not just CI itself but projects can get benefits from alignment on this level.

Patch-set: 2