Update patch set 1
Patch Set 1:

> Are the 'daily' tests appropriate for verify/merge jobs? 
It is not but CI_LOOP doesn't really set test scope to daily.

 > Maybe we should only set that to the merge one?

The tier sets what the test should be and it is healthcheck in the case of verify and merge jobs.

This change is a temporary workaround to ensure we can still run tests but with the supported CI_LOOP vars. I already asked Cedric to have a discussion about it and we will have it today.

Patch-set: 1