Update patch set 1
Patch Set 1:

No, it doesn't depend on it (it would only stop building the container).

As PTL, I take the responsability to go forward and to remove the container. https://gerrit.opnfv.org/gerrit/#/c/54709/ is already well rated and the former Dockerfile could be reused externally for any purpose. From an OPNFV Functest POV, the container doesn't meet the Docker slicing approach we are promoting and it's unused in OPNFV.

Patch-set: 1