jjb: xci: Only set default scenario if no scenario is set
Previously, the code was already setting the default scenario if we had
changes outside the xci/scenario directory. However, this is somewhat
problematic since scenarios may touch other files which do not necessary
affect other scenarios such as anything in xci/playbooks etc. As such,
lets only set a default scenario if no scenario was set up to that
point. However, since installers and scenarios need to match, we also
need to consider installer changes because we fallback to the default
scenario. Finally, since we are filtering results with 'uniq' we can
simply drop the bash regexps prior to adding the scenario to the array.

Change-Id: Ie961071b3e93d04a860bb6655513654208b1a9d4
Signed-off-by: Markos Chandras <mchandras@suse.de>