Move DEPLOY_SCENARIO out of testapi-parameters
These have been moved back into installer-params.yml and added to the
daisy jobs where 'testapi-parameters' is being used.

Including this in the list of parameters would cause duplication as most
installers already include 'DEPLOY_SCENARIO' as a parameter. The
'DEPLOY_SCENARIO' parameter is also sometimes defined through a macro,
which could cause second declaration of DEPLOY_SCENARIO (defaulting to
'os-nosdn-nofeature-ha') to always be used.

Documentation has been added to the parameter to explain the related
macros and how they should be used.

Change-Id: I9aef031dbcf523ae2db15695ce294bd85025530d
Signed-off-by: Trevor Bramwell <tbramwell@linuxfoundation.org>