Update patch set 1
Patch Set 1:

Would it be a problem to keep Danube for a little bit longer for Dovetail? The next release of Dovetail/OVP will be based on Fraser, but we might need to do some updates to the documentation of the Danube-based first release.

Patch-set: 1
Reviewer: Gerrit User 585 <585@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=0