jjb: xci: xci-run-functest: Escape the 'var' variable
Commit 33073c5a3426c1867411db446bc9e5d84d2a7747 ("jjb: xci:
xci-run-functest.sh: Ensure env vars are quoted during reload") added
support for reloading the OPNFV host environment in a safe way but that
didn't work because the internal '$var' variable was not escaped so it
was evaluated on the local host instead of the remote one. This fixes
the problem by properly escaping the variable.

Change-Id: I7c90ae18f77aca076579f70c7c0e9a72aaaad7c8
Signed-off-by: Markos Chandras <mchandras@suse.de>