Verify Tags from Release File
Updates the releases verify job to check release files for the following:

 - commits exist
 - commits are on stable-branch

If the repo has already been tagged the job will not exit, but will warn
that the commit doesn't exist on the correct branch. This is because
tags should never be removed.

This adds an additional script 'repos.py' that provides an quick
interface for pulling out information from release files, along with the
script 'release-status.sh' for checking the status of tags.

NOTE: Disables the branch creation script until it can be reworked.

Change-Id: I498bc74f20aa50d2bd321771f20a77905b246399
Signed-off-by: Trevor Bramwell <tbramwell@linuxfoundation.org>