Create and Verify Tags from Release File
Includes scripts to verify and create tags that rely on it, along with a
script in extra info from the release files.

Running ./releases/scripts/verify-tags.sh first will generate the file
'repos.txt' which is read by ./releases/scripts/create-tags.sh.

This file list repos and the refs in them where the tag should be
created. Currently the tag is hardcoded to 'opnfv-6.0.0', and will not
work for projects which have accidentally created the tag on the master
branch.

TODO: The bash scripts should be moved under jjb/releng as they are
for the merge job.
TODO: Break out renaming of armband to another commit.

Change-Id: I498bc74f20aa50d2bd321771f20a77905b246399
Signed-off-by: Trevor Bramwell <tbramwell@linuxfoundation.org>